슬학생_back/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Pillow==8.3.1
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
